fix: PR #123 코드 리뷰 피드백 반영
Repository 인터페이스 개선:
- saveAndFlush → saveIfNotExists 메서드명 변경 (멱등성 동작 명확화)
- findByIdWithCache → findById 통합 (캐시 세부사항을 구현체에 숨김)

Controller 유효성 검증 추가:
- ProductApiSpec, OrderApiSpec에 @Min/@Max 어노테이션 추가
- page 파라미터에 @Min(0) 적용
- size 파라미터에 @Min(1) @Max(100) 적용
- Controller에서 validation 어노테이션 제거 (인터페이스에서만 정의)

코드 품질 개선:
- ProductFacade에 DEFAULT_PAGE, DEFAULT_SIZE 상수 추출
- Product.increaseLikeCount(int) → increaseLikeCount(long) 타입 일관성

테스트 코드 개선:
- LocalDateTime.now() → 고정 시간 상수 사용 (테스트 재현성 보장)
- 테스트에서 (int) cast 제거 (likeCount 타입 일관성)

+++ apps/commerce-api/src/test/java/com/loopers/interfaces/api/product/ProductApiE2ETest.java
@@ -300,7 +300,7 @@ class ProductApiE2ETest {
   private Product saveProduct(String name, long price, long stock, Long brandId, long likeCount) {
     Product product = Product.of(name, Money.of(price), name + " 상세", Stock.of(stock), brandId);
     if (likeCount > 0) {
-      product.increaseLikeCount((int) likeCount);
+      product.increaseLikeCount(likeCount);
     }
     return productJpaRepository.save(product);
   }
